Backend/requirements.txt
composio
langchain
langchain-openai
langchain-mcp-adapters
python-dotenv
flask
flask-cors
google-api-python-client
google-auth
google-auth-oauthlib
